Adding additional project scaffolding.
Moving away from having a single file. I really don't know what I was
thinking here.

I also hardwired the item with the data structure I was using and
storing the items in a single file along with the object model.
Separating them into their respective files.

Going to need to do like I did with one of my recent projects where I
use pandas maybe to create a dataframe to read in the csv in an easy way
to iterate through to create the item objects.

I also have to go through and play the game a bit more and fumble
through some mixtures to see what other mixtures are possible. I have
reformated the combinates.csv file to be the result first so that it
sorts better. I have no mixtures where cards are involved though for
instance.